feat: 將auth更新成stack auth

+++ src/components/service/ReportList.tsx
@@ -2,6 +2,8 @@
 import React, { useCallback, useEffect, useState } from 'react';
 import { Trash2, ExternalLink, FileX, MessageSquareX } from 'lucide-react';
 import axios from 'axios';
+import { useUser } from '@stackframe/stack';
+
 import {
   Table,
   TableBody,
@@ -34,8 +36,6 @@ import {
   AlertDialogTrigger,
 } from '@/components/ui/alert-dialog';
 
-import { useAuth } from '@/lib/supabase/supbaseContext';
-
 interface ReportListProps {
   serviceId: string;
 }
@@ -46,13 +46,14 @@ const ReportList: React.FC<ReportListProps> = ({ serviceId }) => {
   const [selectedReports, setSelectedReports] = useState<string[]>([]);
   const [deletingItemId, setDeletingItemId] = useState<string | null>(null);
   const { toast } = useToast();
-  const { getAccessToken } = useAuth();
+  const user = useUser();
 
   const fetchReports = useCallback(async () => {
     if (!serviceId) return;
     setIsLoading(true);
     try {
-      const accessToken = await getAccessToken();
+      if (!user) return;
+      const { accessToken } = await user.getAuthJson();
 
       const response = await axios.get(`/api/service/${serviceId}/reports`, {
         headers: { Authorization: `Bearer ${accessToken}` },
@@ -72,7 +73,9 @@ const ReportList: React.FC<ReportListProps> = ({ serviceId }) => {
   const handleDeleteReports = async () => {
     setIsLoading(true);
     try {
-      const accessToken = await getAccessToken();
+      if (!user) return;
+      const { accessToken } = await user.getAuthJson();
+
       await axios.delete(`/api/service/${serviceId}/reports`, {
         headers: { Authorization: `Bearer ${accessToken}` },
         data: { reportIds: selectedReports },
@@ -93,7 +96,9 @@ const ReportList: React.FC<ReportListProps> = ({ serviceId }) => {
     setIsLoading(true);
     setDeletingItemId(report.id);
     try {
-      const accessToken = await getAccessToken();
+      if (!user) return;
+      const { accessToken } = await user.getAuthJson();
+
       if (report.reply?.id || report.thread?.id) {
         await axios.delete(`/api/service/${serviceId}/reports`, {
           headers: { Authorization: `Bearer ${accessToken}` },
